Move everything from vello_api into Vello Common
