+++ sparse_strips/vello_common/src/execute.rs
