+++ sparse_strips/vello_common/src/glyph.rs
@@ -23,6 +23,20 @@ use skrifa::bitmap::{BitmapData, BitmapFormat, BitmapStrikes, Origin};
 pub use vello_api::glyph::*;
 use vello_api::pixmap::Pixmap;
 
+/// Positioned glyph.
+#[derive(Copy, Clone, Default, Debug)]
+pub struct Glyph {
+    /// The font-specific identifier for this glyph.
+    ///
+    /// This ID is specific to the font being used and corresponds to the
+    /// glyph index within that font. It is *not* a Unicode code point.
+    pub id: u32,
+    /// X-offset in run, relative to transform.
+    pub x: f32,
+    /// Y-offset in run, relative to transform.
+    pub y: f32,
+}
+
 /// A type of glyph.
 #[derive(Debug)]
 pub enum GlyphType<'a> {
