+++ sparse_strips/vello_common/src/mask.rs
