+++ sparse_strips/vello_common/src/paint.rs
