+++ sparse_strips/vello_common/src/pixmap.rs
